Creatures are now spawned
