+++ Assets/Scripts/TES/CreatureComponent.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b6da3ff80961ff5489ece374269da0fb
-timeCreated: 1475766237
+guid: 66271927fb4d3874eb1c7d4b8081e793
+timeCreated: 1475839649
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
