+++ Assets/Scripts/TES/NPCComponent.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine;
+
+namespace TESUnity
+{
+    public class NPCComponent : MonoBehaviour
+    {
+    }
+}
