+++ Assets/Scripts/TES/NPCComponent.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: f015af18b2aa5af428c59e00f3ae8016
+timeCreated: 1475839649
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
