+++ Assets/Scripts/TES/Records/Records.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 7184834f6fce64047a2f761b9dc86f3a
+timeCreated: 1475836989
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
